add material

+++ BookStore.BuisneLogic/Models/Authors/AuthorModel.cs
@@ -5,7 +5,7 @@ namespace BookStore.BusinessLogic.Models.Authors
 {
     public class AuthorModel : BaseModel
     {
-        public int PageAmount { get; set; }
+        public long PageAmount { get; set; }
         public ICollection<AuthorModelItem> Items { get; set; }
     }
 }
